Remove ngrok from development workflow
boost cli scripts for add-admin, create-channel, list-channels, and
list-numbers accept a flag '-u' which changes the default url used
from production to whatever url is provided.

Previously, this url was always appended with "https" necessitating the use of a
https proxy (ngrok)

This now changes -u require that the provided url include the protocol. In
development we can use "http://127.0.0.1:3000"

If no url the same results happen as before: boost uses the production server with https.